+++ CHANGELOG.md
@@ -152,6 +152,7 @@ Modifications can only be done via builders.
 - Fix an issue with hidden scopes in graph print outs
 - Fix aliases did not work accross multiple search scenarios
 - Fix no error was thrown when nodes were added to every scope without checking
+- Fix no error was thrown when scopes were added to every scope without checking
 
 ### Removed